+++ bin/install-libraries.sh
@@ -23,32 +23,32 @@ checkAlreadyInstalled
 
 installLibrary flux-authentication-backend https://github.com/fluxfw/flux-authentication-backend/archive/refs/tags/v2023-04-20-2.tar.gz
 
-installLibrary flux-button-group https://github.com/fluxfw/flux-button-group/archive/refs/tags/v2023-05-23-1.tar.gz
+installLibrary flux-button-group https://github.com/fluxfw/flux-button-group/archive/refs/tags/v2023-05-30-1.tar.gz
 
-installLibrary flux-color-scheme https://github.com/fluxfw/flux-color-scheme/archive/refs/tags/v2023-04-24-1.tar.gz
+installLibrary flux-color-scheme https://github.com/fluxfw/flux-color-scheme/archive/refs/tags/v2023-05-30-1.tar.gz
 
 installLibrary flux-config-api https://github.com/fluxfw/flux-config-api/archive/refs/tags/v2023-04-12-1.tar.gz
 
-installLibrary flux-css-api https://github.com/fluxfw/flux-css-api/archive/refs/tags/v2023-04-11-1.tar.gz
+installLibrary flux-css-api https://github.com/fluxfw/flux-css-api/archive/refs/tags/v2023-05-30-1.tar.gz
 
-installLibrary flux-form https://github.com/fluxfw/flux-form/archive/refs/tags/v2023-05-25-1.tar.gz
-
-installLibrary flux-format https://github.com/fluxfw/flux-format/archive/refs/tags/v2023-05-25-1.tar.gz
+installLibrary flux-form https://github.com/fluxfw/flux-form/archive/refs/tags/v2023-05-30-1.tar.gz
 
 installLibrary flux-http-api https://github.com/fluxfw/flux-http-api/archive/refs/tags/v2023-04-20-1.tar.gz
 
-installLibrary flux-loading-spinner https://github.com/fluxfw/flux-loading-spinner/archive/refs/tags/v2023-03-29-1.tar.gz
+installLibrary flux-loading-spinner https://github.com/fluxfw/flux-loading-spinner/archive/refs/tags/v2023-05-30-1.tar.gz
 
 installLibrary flux-mongo-db-connector https://github.com/fluxfw/flux-mongo-db-connector/archive/refs/tags/v2023-04-20-2.tar.gz
 
-installLibrary flux-overlay https://github.com/fluxfw/flux-overlay/archive/refs/tags/v2023-05-25-1.tar.gz
+installLibrary flux-overlay https://github.com/fluxfw/flux-overlay/archive/refs/tags/v2023-05-30-1.tar.gz
 
-installLibrary flux-pwa-api https://github.com/fluxfw/flux-pwa-api/archive/refs/tags/v2023-05-01-1.tar.gz
+installLibrary flux-pwa-api https://github.com/fluxfw/flux-pwa-api/archive/refs/tags/v2023-05-30-1.tar.gz
 
 installLibrary flux-pwa-generator https://github.com/fluxfw/flux-pwa-generator/archive/refs/tags/v2023-04-24-1.tar.gz
 
 installLibrary flux-shutdown-handler https://github.com/fluxfw/flux-shutdown-handler/archive/refs/tags/v2023-03-16-1.tar.gz
 
-installLibrary flux-table https://github.com/fluxfw/flux-table/archive/refs/tags/v2023-05-25-1.tar.gz
+installLibrary flux-table https://github.com/fluxfw/flux-table/archive/refs/tags/v2023-05-30-1.tar.gz
+
+installLibrary flux-value-format https://github.com/fluxfw/flux-value-format/archive/refs/tags/v2023-05-30-1.tar.gz
 
 installLibrary mime-db https://registry.npmjs.org/mime-db/-/mime-db-1.52.0.tgz