+++ src/FieldType/ColorFieldType.mjs
@@ -1,6 +1,6 @@
 import { FIELD_TYPE_COLOR } from "./FIELD_TYPE.mjs";
-import { FORMAT_TYPE_COLOR } from "../../../flux-format/src/FORMAT_TYPE.mjs";
 import { INPUT_TYPE_COLOR } from "../../../flux-form/src/INPUT_TYPE.mjs";
+import { VALUE_FORMAT_TYPE_COLOR } from "../../../flux-value-format/src/VALUE_FORMAT_TYPE.mjs";
 
 /** @typedef {import("../Field/Field.mjs").Field} Field */
 /** @typedef {import("./FieldType.mjs").FieldType} FieldType */
@@ -46,8 +46,8 @@ export class ColorFieldType {
      * @param {Field} field
      * @returns {Promise<string>}
      */
-    async getFormatType(field) {
-        return FORMAT_TYPE_COLOR;
+    async getValueFormatType(field) {
+        return VALUE_FORMAT_TYPE_COLOR;
     }
 
     /**