+++ src/FieldType/EmailFieldType.mjs
@@ -1,5 +1,5 @@
 import { FIELD_TYPE_EMAIL } from "./FIELD_TYPE.mjs";
-import { FORMAT_TYPE_EMAIL } from "../../../flux-format/src/FORMAT_TYPE.mjs";
+import { VALUE_FORMAT_TYPE_EMAIL } from "../../../flux-value-format/src/VALUE_FORMAT_TYPE.mjs";
 import { INPUT_TYPE_EMAIL, INPUT_TYPE_TEXT } from "../../../flux-form/src/INPUT_TYPE.mjs";
 
 /** @typedef {import("../Field/Field.mjs").Field} Field */
@@ -58,8 +58,8 @@ export class EmailFieldType {
      * @param {Field} field
      * @returns {Promise<string>}
      */
-    async getFormatType(field) {
-        return FORMAT_TYPE_EMAIL;
+    async getValueFormatType(field) {
+        return VALUE_FORMAT_TYPE_EMAIL;
     }
 
     /**