+++ src/FieldType/MultilineTextFieldType.mjs
@@ -1,5 +1,5 @@
 import { FIELD_TYPE_MULTILINE_TEXT } from "./FIELD_TYPE.mjs";
-import { FORMAT_TYPE_TEXT } from "../../../flux-format/src/FORMAT_TYPE.mjs";
+import { VALUE_FORMAT_TYPE_TEXT } from "../../../flux-value-format/src/VALUE_FORMAT_TYPE.mjs";
 import { INPUT_TYPE_TEXT, INPUT_TYPE_TEXTAREA } from "../../../flux-form/src/INPUT_TYPE.mjs";
 
 /** @typedef {import("../Field/Field.mjs").Field} Field */
@@ -56,8 +56,8 @@ export class MultilineTextFieldType {
      * @param {Field} field
      * @returns {Promise<string>}
      */
-    async getFormatType(field) {
-        return FORMAT_TYPE_TEXT;
+    async getValueFormatType(field) {
+        return VALUE_FORMAT_TYPE_TEXT;
     }
 
     /**