+++ src/FieldType/RegularExpressionFieldType.mjs
@@ -1,7 +1,7 @@
 import { ADDITIONAL_VALIDATION_TYPE_REGULAR_EXPRESSION } from "../../../flux-form/src/ADDITIONAL_VALIDATION_TYPE.mjs";
 import { FIELD_TYPE_REGULAR_EXPRESSION } from "./FIELD_TYPE.mjs";
-import { FORMAT_TYPE_TEXT } from "../../../flux-format/src/FORMAT_TYPE.mjs";
 import { INPUT_TYPE_TEXT } from "../../../flux-form/src/INPUT_TYPE.mjs";
+import { VALUE_FORMAT_TYPE_TEXT } from "../../../flux-value-format/src/VALUE_FORMAT_TYPE.mjs";
 
 /** @typedef {import("../Field/Field.mjs").Field} Field */
 /** @typedef {import("./FieldType.mjs").FieldType} FieldType */
@@ -69,8 +69,8 @@ export class RegularExpressionFieldType {
      * @param {Field} field
      * @returns {Promise<string>}
      */
-    async getFormatType(field) {
-        return FORMAT_TYPE_TEXT;
+    async getValueFormatType(field) {
+        return VALUE_FORMAT_TYPE_TEXT;
     }
 
     /**