+++ src/UI/FluxFieldValueStorageUI.mjs
@@ -1,8 +1,8 @@
 import { COLOR_SCHEME_LIGHT } from "./Libs/flux-color-scheme/src/ColorScheme/COLOR_SCHEME.mjs";
 import { flux_css_api } from "./Libs/flux-css-api/src/FluxCssApi.mjs";
-import { FORMAT_TYPE_FLUX_TABLE_ACTIONS } from "./Libs/flux-table/src/FORMAT_TYPE.mjs";
 import { HttpClientRequest } from "./Libs/flux-http-api/src/Client/HttpClientRequest.mjs";
 import { INPUT_TYPE_ENTRIES } from "./Libs/flux-form/src/INPUT_TYPE.mjs";
+import { VALUE_FORMAT_TYPE_FLUX_TABLE_ACTIONS } from "./Libs/flux-table/src/VALUE_FORMAT_TYPE.mjs";
 import { METHOD_DELETE, METHOD_POST, METHOD_PUT } from "./Libs/flux-http-api/src/Method/METHOD.mjs";
 import { ROW_ACTION_UPDATE_TYPE_DISABLE_ON_FIRST, ROW_ACTION_UPDATE_TYPE_DISABLE_ON_LAST } from "./Libs/flux-table/src/ROW_ACTION_UPDATE_TYPE.mjs";
 
@@ -10,18 +10,17 @@ import { ROW_ACTION_UPDATE_TYPE_DISABLE_ON_FIRST, ROW_ACTION_UPDATE_TYPE_DISABLE
 /** @typedef {import("./Libs/flux-button-group/src/FluxButtonGroupElement.mjs").FluxButtonGroupElement} FluxButtonGroupElement */
 /** @typedef {import("./Libs/flux-color-scheme/src/FluxColorScheme.mjs").FluxColorScheme} FluxColorScheme */
 /** @typedef {import("./Libs/flux-form/src/FluxFormElement.mjs").FluxFormElement} FluxFormElement */
-/** @typedef {import("./Libs/flux-format/src/FluxFormat.mjs").FluxFormat} FluxFormat */
+/** @typedef {import("./Libs/flux-value-format/src/FluxValueFormat.mjs").FluxValueFormat} FluxValueFormat */
 /** @typedef {import("./Libs/flux-http-api/src/FluxHttpApi.mjs").FluxHttpApi} FluxHttpApi */
 /** @typedef {import("./Libs/flux-pwa-api/src/FluxPwaApi.mjs").FluxPwaApi} FluxPwaApi */
 /** @typedef {import("./Libs/flux-form/src/InputValue.mjs").InputValue} InputValue */
 /** @typedef {import("../Value/ValueTable.mjs").ValueTable} ValueTable */
 
-flux_css_api.adopt(
-    document,
-    await flux_css_api.import(
-        `${import.meta.url.substring(0, import.meta.url.lastIndexOf("/"))}/Layout/style.css`
-    )
+const layout_css = await flux_css_api.import(
+    `${import.meta.url.substring(0, import.meta.url.lastIndexOf("/"))}/Layout/style.css`
 );
+
+document.adoptedStyleSheets.push(layout_css);
 
 export class FluxFieldValueStorageUI {
     /**
@@ -33,9 +32,9 @@ export class FluxFieldValueStorageUI {
      */
     #flux_color_scheme = null;
     /**
-     * @type {FluxFormat | null}
+     * @type {FluxValueFormat | null}
      */
-    #flux_format = null;
+    #flux_value_format = null;
     /**
      * @type {FluxHttpApi | null}
      */
@@ -809,7 +808,7 @@ export class FluxFieldValueStorageUI {
         }
 
         const flux_table_element = await (await import("./Libs/flux-table/src/FluxTableElement.mjs")).FluxTableElement.newWithData(
-            await this.#getFluxFormat(),
+            await this.#getFluxValueFormat(),
             [
                 {
                     action: async () => {
@@ -839,7 +838,7 @@ export class FluxFieldValueStorageUI {
                 {
                     key: "actions",
                     label: "Actions",
-                    "format-type": FORMAT_TYPE_FLUX_TABLE_ACTIONS
+                    "value-format-type": VALUE_FORMAT_TYPE_FLUX_TABLE_ACTIONS
                 }
             ],
             "name",
@@ -954,12 +953,12 @@ export class FluxFieldValueStorageUI {
     }
 
     /**
-     * @returns {Promise<FluxFormat>}
+     * @returns {Promise<FluxValueFormat>}
      */
-    async #getFluxFormat() {
-        this.#flux_format ??= (await import("./Libs/flux-format/src/FluxFormat.mjs")).FluxFormat.new();
+    async #getFluxValueFormat() {
+        this.#flux_value_format ??= (await import("./Libs/flux-value-format/src/FluxValueFormat.mjs")).FluxValueFormat.new();
 
-        return this.#flux_format;
+        return this.#flux_value_format;
     }
 
     /**
@@ -1046,7 +1045,7 @@ export class FluxFieldValueStorageUI {
         }
 
         this.#value_element.appendChild(await (await import("./Libs/flux-table/src/FluxTableElement.mjs")).FluxTableElement.newWithData(
-            await this.#getFluxFormat(),
+            await this.#getFluxValueFormat(),
             [
                 {
                     action: () => {
@@ -1078,7 +1077,7 @@ export class FluxFieldValueStorageUI {
                 {
                     key: "actions",
                     label: "Actions",
-                    "format-type": FORMAT_TYPE_FLUX_TABLE_ACTIONS
+                    "value-format-type": VALUE_FORMAT_TYPE_FLUX_TABLE_ACTIONS
                 }
             ]) ?? null,
             "name",